Support parsing the minimum long value
Previously, when parsing integers, the `ValueParser` separated out the
sign and tried to parse them as positive longs. That did not work for
`Long.MIN_VALUE` as that is `-9223372036854775808` but the maximum
positive long is `9223372036854775807`. The reason for separating out
the sign that way was in case the original sign (specifically a `+`)
needed to be preserved. However, it is not clear if that is needed so
there is no point in trying to support that yet.

This change removes `UnaryOperator` and just includes the sign in the
string that is converted to the number, for all numbers.

Bug: 354633349
Test: ./gradlew
Change-Id: Ib6b3349359c1c52626304f4380e28ed8966c357c