Fix Kotlin nested annotation testing
Previously, Kotlin nested annotation tests used a `@` prefix on the
nested annotation just like Java. However, Kotlin does not support that
syntax. Instead it just treats nested annotations as if they were
constructed like a normal class. It also used `Class<*>` instead of
`KClass<*>` in the definition of an annotation property that can accept
a class object.

Fixing that revealed an issue with the `UAnnotationItem`. It could not
actually handle nested annotations. It does have a `UAnnotation` branch
in the `when (value) {...}` block but that is never actually taken.
Instead the `UCallExpression` branch is taken (as Kotlin treats
annotations as normal classes to be constructed). Unfortunately, that
only handles arrays and so it drops out of the `when` without appending
anything.

Prior to fixing the test it ended up going through the `else` branch of
the `when` as an unknown (or invalid) expression.

This change moves the `else` branch of the `when` block after the block
so that any `value` that is not handled in the `when` block will append
something. The branches that do append something return to avoid that.
It also fixes the tests to match the current output.

Once the `Value` model has been fully implemented this code will be
removed in its entirety.

Bug: 354633349
Test: ./gradlew
Change-Id: Iedef5ac0aba14784b945bffdfae8cf27b9be842d