Move using separator as the token out of scanForEndOfToken(...)
Previously, the code to handle the case when a separator character was
the first, and so only, character in the token was part of the
`scanForEndOfToken(...)`. However, that differs from the rest of the
code in the `scanForEndOfToken(...)` method which handles characters
anywhere in the token.

This change pulls that code up into `getToken(...)` to make
`scanForEndOfToken(...)` more consistent and allow it to be refactored
to improve handling of open and close characters like `<` and `>`.

Bug: 354633349
Test: ./gradlew
Change-Id: I7c470c1ef2336b0d7615a8b91f553fa47ce743f9