Replace parenIsSep with TokenPurpose
Previously, `Tokenizer` used `parenIsSep` to determine whether to treat
`(` and `)` as separators or to balance them out as part of the token.
This change replaces that `Boolean` with a `TokenPurpose` that replaces
the `true` and `false` with more meaningful options. It also abstracts
it away from just being about parentheses.

Needed for a follow-up change that will use it to balance `{` and `}`
too.

Bug: 354633349
Test: ./gradlew
Change-Id: Ia8136215fe7e27251456fce4c3a094ff76a2f6a8