Improve Turbine support for ConstantFieldValue
In order for Turbine to support `ConstantFieldValue`s it needs to take
a `ConstVarName` expression and resolve it to a `FieldInfo` using the
`TurbineFieldResolver`. The meaning of the `ConstVarName` depends on
the class it is in, the imports that are available, etc.. All of that
is handled by the `TurbineFieldResolver` of the class containing the
`ConstVarName` expression.

The `TurbineFieldResolver` is stored in the `TurbineValue` for use by
the `TurbineValueFactory` which is responsible for converting the
`ConstVarName` into a `FieldInfo` into a `ConstantFieldValue`.

Previously, the `TurbineFieldResolver` was only passed to the
`TurbineValue` for `MethodItem.defaultValue`. This change supplies it
for `TurbineValue`s used for non-type use `AnnotationAttribute.value`
and `FieldItem.initialValue`s. The latter is not strictly necessary as
currently `FieldItem.initialValue` does not reference other fields but
it helps with the goal is to always supply it when constructing a
`TurbineValue`.

Bug: 354633349
Test: ./gradlew
Change-Id: I363f5ee6b002f3a41b415d88df36bc1b12f8eea5