Add FieldItem.constantValue
It is difficult to determine the difference between
`legacyInitialValue(true)` and `legacyInitialValue(false)` as their
standard behavior is defined by how it behaves in Psi and the Psi
implementation (in `PsiFieldValue`) uses undocumented methods.

Previously, it was assumed that the only difference between them was
the `Value` itself. However, the former only returns a value at all for
fields which are `static final`.

This change adds a `FieldItem.constantValue` to replicate that
behavior and replaces all existing uses of `FieldItem.literalValue`
with it as they previously used `legacyInitialValue(true)`.

Bug: 354633349
Test: ./gradlew
Change-Id: I6a05e764e820f94270790ec556ec1d3083657df5