Fix non-constant values in FieldItem.writeValueWithSemicolon(...)
Previously, `FieldItem.writeValueWithSemicolon(...)` and
`JavaStubWriter` shared the work on deciding whether to write a value.
The resulting code was complicated and incorrect with the following
issues:
1. It would include values on fields which were not `static final`.
   While not an error it was unnecessary.
2. Even worse it would use a default value for a `static final`
   primitive field that was different from the actual value (which was
   computed at runtime) but would be inlined by javac.
3. It would sometimes not provide an initializer at all which would
   fail to compile.

This simplifies both by having `FieldItem.writeValueWithSemicolon(...)`
just write out the value it was given, if possible, and return `true`
if a value was written and `false` otherwise. It also takes an optional
`nonConstantExpressionProvider` which, if provided, will supply a
non-constant expression for initializing `static final` fields that do
not have a value and require an initializer.

`JavaStubWriter` supplies a `nonConstantExpressionProvider` so that the
resulting stubs will compile, while `SignatureWriter` does not because
it does not care if a `static final` field has a value.

Bug: 407951018
Test: ./gradlew
Change-Id: Ia1307919782663de3551f1d6aba90273ad234a75