Treat ConstantFieldValue as its ConstantValue
Previously, a reference to a constant field would just use the field's
`constantValue`, even if it was a different type to where the field was
used. e.g. in `long value = INT_CONSTANT`, the value would be the value
of `INT_CONSTANT`. So, if `INT_CONSTANT` was 37 then the value be an
`IntValue` of `37`. However, that is incorrect. It should behave
exactly as if the field reference was replaced with its value, e.g. the
above expression should behave as `long value = 37`. That would produce
a `LongValue` that was specified as an int.

This change fixes that issue.

Bug: 354633349
Test: ./gradlew
Change-Id: I41ac875f8c08ee43057fa31fe29085e0bb0637ef