Evaluate binary expressions in AnnotationItem.toSource()
Previously, Psi would output `PsiBinaryExpression` and
`UBinaryExpression` in `toSource()`. That meant that signature files
were coupled to how an annotation's values were specified in the source
and complicates signature file parsing.

This change evaluates the binary expressions instead of outputting
them. There is one exception which is the floating point special
values, i.e. positive and negative infinities and NaN. They are still
represented as a division by 0.

A signature file test using complex expressions was removed altogether.
It was never a realistic test. A couple of other tests were affected
and fixed.

Bug: 410756171
Test: ./gradlew
      # Checked impact on Android and AndroidX and it is minimal.
Change-Id: I2218eb0d40795a82dd50416223f2f0df9ed29b54