Remove requireConstant from legacyInitialValue(...)
Previously, most of the existing uses of `legacyInitialValue(...)`
passed `true` (either explicitly or implicitly) for the
`requireConstant` parameter. Only tests passed `false`.

This change removes most test code that calls
`legacyInitialValue(false)`. As `initialValue` was intended as the
replacement for `legacyInitialValue(false)` that will be removed in a
follow-up change so this also replaces any remaining uses of that with
`constantValue` to be consistent.

There was one slight issue with `PsiFieldItemTest` which tested the
behavior of `legacyInitialValue(false)` but not
`legacyInitialValue(true)` and did not test `initialValue` or
`constantValue` either. This change makes that test consistent with
the others by testing `legacyInitialValue()` and `constantValue`
side-by-side.

Note, `FieldValue.initialValue(...)` still has a `requireConstant` as
there are some Psi specific uses of that which will be cleaned up
in a follow-up change.

Bug: 354633349
Test: ./gradlew
Change-Id: I601fb443d19d23055631bca83c7163554755a7f7