Only create ConstantValues for fields which could be constant
Previously, `PsiFieldItem` would create a `constantValueProvider` for
fields which could not have a compile-time constant value, e.g. Java
fields which are not `static` and `final`, or Kotlin fields which are
not `const`. `ApiFile` would also create a `constantValueProvider` for
fields which are not `static` and `final` (which all constant fields in
the signature file must be). Ditto for Turbine.

This change makes `PsiFieldItem` only create a `constantValueProvider`
for fields which could have a constant value. It does the same for
`TurbineClassBuilder` and `ApiFile`.

As signature files should only have a field initializer for a constant
field it will report an issue if it finds a field initializer for a
non-`static`, non-`final` field.

The latter revealed a few issues with some tests which were fixed.

Bug: 354633349
Test: ./gradlew
Change-Id: I9c79688ccef72a8f680a15c7311d175d4532495c