Improve formatting of range limits
Change https://r.android.com/3592332 inadvertently changed the
formatting of range limits in the Javadoc. Prior to that the
formatting was inconsistent as it just used the source representation.
However, when using a constant field it did include the name of the
field while the current formatting does not.

This change adds tests and improves the formatting as follows:
* Integral values will be output as an integer unless it is too big
  in which case it will output it as a long, i.e. with `L` suffix.
* Similarly, floating point values will be output as a float (with
  'f` suffix) unless it is too big in which case it will be output as
  a double.
* Field references will be output as an `@link`.

Bug: 354633349
Test: ./gradlew
Change-Id: I9b6a6e6630e0e30171e4f8ce88f6ed909a941b8f