Remove legacyFieldValue and associated code
Previously, `legacyFieldValue` was only really used in the
`CommonParameterizedFieldValueTest.testLegacyValue()` method so this
removes that completely and all baselined failures in it.

After stopping passing `FieldValue` to the `DefaultFieldItem`
constructor the `ApiFile`'s `fieldValue` local variable was unused so
it was removed, along with the now unused `TextFieldValue`.

Similarly, `FieldValue.snapshot()` was no longer used so that was
removed too.

Both Turbine and Psi still make use of their `FieldValue` instance but
they will be cleaned up in a follow-up changes.

Bug: 354633349
Test: ./gradlew
Change-Id: I29ed581a104b87a3c322bb7333a9e8f85ada49a6