Remove PsiFieldValue and FieldValue
Previously, Turbine used `PsiFieldValue` in the `isInitialValueNonNull`
parameter passed to `TypeItemFactory.getFieldType(...)`. However, that
does not need to create a `PsiFieldValue` it just needs to check
whether, if one was created, would it be non-null. So, this change
refactors the creation code to perform that check and removes the
`PsiFieldValue` and `FieldValue` as that was no longer used.

Bug: 354633349
Test: ./gradlew
Change-Id: I7a6cee7cd1597b318ce9b6b63effab76fd3a15b2